LIVE-1575: rename Headline.tsx to index.tsx
